DAO-188 Add PolicySelect page
