+++ src/components/policies/index.tsx
@@ -1,1 +1,2 @@
 export { default } from './policies';
+export type { Filter } from './policies';
