+++ src/pages/policy-select/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './policy-select';
